マージ

+++ Test/Flashlight.cpp
@@ -29,7 +29,7 @@ void Flashlight::Initialize(const Transform* parentTransform, const Camera* pare
 	subLightPower_ = 0.1f;
 	isLighting_ = true;
 
-	collider_ = std::make_shared<ConeCollider>(CollisionCategory::PLAYER, static_cast<uint32_t>(CollisionCategory::LIGHT | CollisionCategory::ENEMY | CollisionCategory::ITEM), transform_.translate, fovAngle_, lightRange_, transform_.rotate);
+	collider_ = std::make_shared<ConeCollider>(CollisionCategory::PLAYER, (CollisionCategory::LIGHT | CollisionCategory::ENEMY | CollisionCategory::ITEM), transform_.translate, fovAngle_, lightRange_, transform_.rotate);
 }
 
 void Flashlight::Update()
